Fix API call for new Gandi API
I tried using this plugin via the Debian package repositories but it
kept erroring. Further investigation showed that the base URL used in
this script was incorrect, as Gandi have made upstream changes.

This commit changes the URL to the new one and amends the authentication
system to match. I was able to renew my certificates using this amended
script.

Fixes #19.